//解决播放器暂停状态下切换全屏黑屏的问题

+++ app/src/main/java/com/leowong/project/eyepetizer/media/IMediaPlayerControl.java
@@ -37,7 +37,7 @@ public interface IMediaPlayerControl {
 
     long getCurrentPosition();
 
-    void seekTo(int pos);
+    void seekTo(long pos);
 
     void play(Uri videoUri, Long position);
